+++ test/modules/attestation/Attestation.test.ts
@@ -16,9 +16,9 @@
  */
 
 import { ChainId, DeleteAccountPropertyParams, GetAccountPropertiesParams, SetAccountPropertyParams } from "@somedotone/ardor-ts";
-import { Attestation, attestation, CreateAttestationUncheckedParams, CreateIntermediateAttestationParams, CreateLeafAttestationParams, CreateRootAttestationParams, EntityType, Error, ErrorCode, RevokeAttestationUncheckedParams, RevokeIntermediateAttestationParams, RevokeLeafAttestationParams, RevokeRootAttestationParams, State, UpdateIntermediateAttestationParams, UpdateLeafAttestationParams, UpdateRootAttestationParams } from "../src/index";
-import config from "./config";
-import RequestMock from "./mocks/RequestMock";
+import { Attestation, attestation, CreateAttestationUncheckedParams, CreateIntermediateAttestationParams, CreateLeafAttestationParams, CreateRootAttestationParams, EntityType, Error, ErrorCode, RevokeAttestationUncheckedParams, RevokeIntermediateAttestationParams, RevokeLeafAttestationParams, RevokeRootAttestationParams, State, UpdateIntermediateAttestationParams, UpdateLeafAttestationParams, UpdateRootAttestationParams } from "../../../src/index";
+import config from "../../config";
+import RequestMock from "../../mocks/RequestMock";
 
 
 if (config.test.attestationModule.runTests) {
